requirements.txt
PyPDF2==2.0
Pillow==9.1.1
virustotal-api==1.1.11
python-nmap==0.7.1
XlsxWriter==3.0.3
pprintpp==0.4.0
